@@ -19,32 +19,32 @@
     "ci": "run-s build:all lint:all test:all git_is_clean"
   },
   "devDependencies": {
-    "@essex/eslint-config": "^15.0.4",
-    "@essex/eslint-plugin": "^15.0.4",
-    "@essex/jest-config": "^15.0.3",
-    "@essex/scripts": "^16.0.4",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@tsconfig/node14": "^1.0.0",
+    "@essex/eslint-config": "^17.0.0",
+    "@essex/eslint-plugin": "^17.0.0",
+    "@essex/jest-config": "^16.0.0",
+    "@essex/scripts": "^17.0.0",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@tsconfig/node14": "^1.0.1",
     "@types/jest": "^26.0.23",
     "@types/lodash": "^4.14.170",
     "@types/prettier": "^2.2.3",
     "@types/testing-library__jest-dom": "^5.9.5",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "eslint": "^7.26.0",
+    "@typescript-eslint/eslint-plugin": "^4.28.0",
+    "@typescript-eslint/parser": "^4.28.0",
+    "eslint": "^7.29.0",
     "eslint-import-resolver-node": "^0.3.4",
     "eslint-plugin-header": "^3.1.1",
-    "eslint-plugin-import": "^2.23.2",
+    "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-jest": "^24.3.6",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.23.2",
+    "eslint-plugin-react": "^7.24.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "husky": "^6.0.0",
     "jest": "^26.6.3",
     "lint-staged": "^11.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.0",
-    "typescript": "^4.2.4"
+    "prettier": "^2.3.1",
+    "typescript": "^4.3.4"
   },
   "workspaces": [
     "packages/*"
@@ -56,7 +56,7 @@
     "react-dom": "17.0.2"
   },
   "dependencies": {
-    "@fluentui/react-hooks": "^8.2.1",
+    "@fluentui/react-hooks": "^8.2.2",
     "lodash": "^4.17.21"
   }
 }
