@@ -6,8 +6,7 @@
     "clean": "essex clean dist/ deploy.zip",
     "start": "nodemon src/index.ts",
     "start:api": "yarn start",
-    "bundle": "run-s install_deploy_deps create_deploy_archive",
-    "build": "run-s compile copy_assets write_deploy_package",
+    "build": "run-s compile copy_assets write_deploy_package install_deploy_deps create_deploy_archive",
     "compile": "ttsc -b tsconfig.json",
     "test": "essex test",
     "create_mock_data": "ts-node createMockData.ts",
@@ -19,45 +18,51 @@
     "write_deploy_package": "ts-node writeDeployPackage.ts"
   },
   "dependencies": {
+    "@graphql-tools/schema": "^7.1.5",
     "@greenlight/schema": "workspace:packages/schema",
+    "apollo-server": "^2.25.2",
+    "apollo-server-core": "^2.25.2",
+    "apollo-server-fastify": "^2.25.2",
     "bcrypt": "^5.0.1",
     "config": "^3.3.6",
     "dataloader": "^2.0.0",
     "detect-libc": "^1.0.3",
-    "fastify": "^3.15.1",
+    "fastify": "^2.15.3",
     "fastify-cors": "^6.0.1",
-    "fastify-jwt": "^2.5.0",
-    "fastify-nodemailer": "^5.0.0",
-    "graphql": "^15.5.0",
+    "graphql": "^15.5.1",
+    "graphql-subscriptions": "^1.2.1",
+    "graphql-tools": "^7.0.5",
+    "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
     "make-dir": "^3.1.0",
-    "mercurius": "^7.9.0",
-    "mercurius-auth": "^1.1.0",
-    "mongodb": "^3.6.8",
+    "mongodb": "^3.6.9",
     "node-fetch": "^2.6.1",
+    "nodemailer": "^6.6.2",
     "nopt": "^5.0.0",
     "npmlog": "^4.1.2",
     "pino": "^6.11.3",
-    "pino-pretty": "^4.8.0",
+    "pino-pretty": "^5.0.2",
     "semver": "^7.3.5",
     "tar": "^6.1.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/preset-typescript": "^7.13.0",
-    "@essex/jest-config": "^15.0.3",
-    "@essex/scripts": "^16.0.4",
-    "@tsconfig/node14": "^1.0.0",
+    "@babel/core": "^7.14.6",
+    "@babel/preset-env": "^7.14.7",
+    "@babel/preset-react": "^7.14.5",
+    "@babel/preset-typescript": "^7.14.5",
+    "@essex/jest-config": "^16.0.0",
+    "@essex/scripts": "^17.0.0",
+    "@tsconfig/node14": "^1.0.1",
     "@types/bcrypt": "^5.0.0",
     "@types/config": "^0.0.38",
     "@types/faker": "^5.5.5",
     "@types/jest": "^26.0.23",
+    "@types/jsonwebtoken": "^8",
     "@types/lodash": "^4.14.170",
-    "@types/mongodb": "^3.6.16",
-    "@types/node": "^14.17.0",
+    "@types/mongodb": "^3.6.18",
+    "@types/node": "^14.17.3",
+    "@types/nodemailer": "^6",
     "@types/pino": "^6.3.8",
     "@types/pino-pretty": "^4.7.0",
     "@types/uuid": "^8.3.0",
@@ -69,7 +74,7 @@
     "tsconfig-paths": "^3.9.0",
     "tsconfig-paths-jest": "^0.0.1",
     "ttypescript": "^1.5.12",
-    "typescript": "^4.2.4",
+    "typescript": "^4.3.4",
     "typescript-transform-paths": "^2.2.4"
   }
 }
