@@ -21,22 +21,22 @@
 		"plop:webapp": "plop"
 	},
 	"dependencies": {
-		"@apollo/client": "^3.3.19",
-		"@fluentui/react": "^8.14.13",
-		"@fluentui/react-hooks": "^8.2.1",
+		"@apollo/client": "^3.3.20",
+		"@fluentui/react": "^8.20.2",
+		"@fluentui/react-hooks": "^8.2.2",
 		"@greenlight/schema": "workspace:packages/schema",
 		"bootstrap": "^5.0.1",
 		"classnames": "^2.3.1",
 		"config": "^3.3.6",
 		"express": "^4.17.1",
-		"formik": "^2.2.7",
-		"graphql": "^15.5.0",
+		"formik": "^2.2.9",
+		"graphql": "^15.5.1",
 		"json2csv": "^5.0.6",
 		"lodash": "^4.17.21",
-		"next": "^10.2.2",
-		"next-auth": "^3.23.2",
+		"next": "^10.2.3",
+		"next-auth": "^3.27.0",
 		"react": "^17.0.2",
-		"react-bootstrap": "^1.6.0",
+		"react-bootstrap": "^1.6.1",
 		"react-dom": "^17.0.2",
 		"react-dom-confetti": "^0.2.0",
 		"react-is": "^17.0.2",
@@ -44,39 +44,38 @@
 		"react-select": "^4.3.1",
 		"react-transition-group": "^4.4.2",
 		"recoil": "^0.3.1",
-		"recoil-persist": "^2.5.0",
+		"recoil-persist": "^2.9.0",
 		"styled-components": "^5.3.0",
 		"subscriptions-transport-ws": "^0.9.19",
 		"yup": "^0.32.9"
 	},
 	"license": "MIT",
 	"devDependencies": {
-		"@essex/scripts": "^16.0.4",
-		"@tsconfig/node14": "^1.0.0",
+		"@essex/scripts": "^17.0.0",
+		"@tsconfig/node14": "^1.0.1",
 		"@types/config": "^0.0.38",
 		"@types/express": "^4.17.11",
 		"@types/json2csv": "^5",
-		"@types/node": "^14.17.0",
+		"@types/node": "^14.17.3",
 		"@types/react": "^17.0.6",
 		"@types/react-dom": "^17.0.5",
 		"@types/react-is": "^17",
 		"@types/react-select": "^4",
 		"@types/react-transition-group": "^4",
-		"@types/recoil": "^0",
 		"@types/styled-components": "^5",
 		"@types/yup": "^0.29.11",
-		"autoprefixer": "^10.2.5",
+		"autoprefixer": "^10.2.6",
 		"babel-plugin-transform-define": "^1.3.2",
 		"nodemon": "^2.0.7",
 		"npm-run-all": "^4.1.5",
 		"plop": "^2.7.4",
-		"postcss": "^8.3.0",
+		"postcss": "^8.3.5",
 		"postcss-flexbugs-fixes": "^5.0.2",
 		"postcss-hover-media-feature": "^1.0.1",
 		"postcss-preset-env": "^6.7.0",
 		"sass": "1.32.13",
 		"shx": "^0.3.3",
 		"ts-node": "^9.1.1",
-		"typescript": "^4.2.4"
+		"typescript": "^4.3.4"
 	}
 }
