@@ -52,13 +52,14 @@ const createAuthLink = () => {
 	return _authLink.concat(httpLink)
 }
 
-const createWSLink = () => {
+const createWebSocketLink = () => {
 	const authorization = createAuthorizationHeader()
 
 	return new WebSocketLink(
 		new SubscriptionClient(process.env.API_SOCKET_URL, {
 			lazy: true,
 			reconnect: true,
+			reconnectionAttempts: 3,
 			connectionParams: async () => {
 				return {
 					headers: {
@@ -72,7 +73,7 @@ const createWSLink = () => {
 
 const createSplitLink = () => {
 	const authLink = createAuthLink()
-	const wsLink = createWSLink()
+	const wsLink = createWebSocketLink()
 
 	return split(
 		({ query }) => {
