@@ -7,7 +7,7 @@ import type ComponentProps from '~types/ComponentProps'
 import NotificationRow from '~ui/NotificationRow'
 import { useAuthUser } from '~hooks/api/useAuth'
 import { useRouter } from 'next/router'
-import { useCurrentUser } from '~hooks/api/useCurrentuser'
+// import { useCurrentUser } from '~hooks/api/useCurrentuser'
 
 interface NotificationPanelBodyProps extends ComponentProps {
 	onClose?: () => void
@@ -16,14 +16,14 @@ interface NotificationPanelBodyProps extends ComponentProps {
 export default function NotificationPanelBody({
 	onClose
 }: NotificationPanelBodyProps): JSX.Element {
-	const { markMention } = useAuthUser()
-	const { currentUser } = useCurrentUser()
+	const { authUser, markMention } = useAuthUser()
+	// const { currentUser } = useCurrentUser()
 
 	const router = useRouter()
 
 	const handleNotificationSelect = async (engagementId, seen) => {
 		if (!seen) {
-			await markMention(currentUser.id, engagementId)
+			await markMention(authUser.user.id, engagementId)
 		}
 		router.push(`${router.pathname}?engagement=${engagementId}`)
 	}
@@ -32,11 +32,11 @@ export default function NotificationPanelBody({
 		<div className={styles.bodyWrapper}>
 			<h3>Notifications</h3>
 
-			{(!currentUser?.mentions || currentUser?.mentions.length === 0) && (
+			{(!authUser.user?.mentions || authUser.user?.mentions.length === 0) && (
 				<div className={styles.noMentions}>You have no notifications. </div>
 			)}
 
-			{currentUser?.mentions?.map((m, i) => (
+			{authUser.user?.mentions?.map((m, i) => (
 				<NotificationRow
 					key={`${m.engagementId}-${i}`}
 					clickCallback={() => handleNotificationSelect(m.engagementId, m.seen)}
