@@ -3,10 +3,12 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 import { gql, useLazyQuery } from '@apollo/client'
-import type { AuthenticationResponse, User } from '@greenlight/schema/lib/client-types'
+import type { User } from '@greenlight/schema/lib/client-types'
+// import type { AuthenticationResponse, User } from '@greenlight/schema/lib/client-types'
 import { useEffect } from 'react'
 import { useRecoilState } from 'recoil'
-import { currentUserState, userAuthState } from '~store'
+import { currentUserState } from '~store'
+// import { currentUserState, userAuthState } from '~store'
 import { CurrentUserFields } from './fragments'
 
 const CURRENT_USER = gql`
@@ -19,26 +21,26 @@ const CURRENT_USER = gql`
 	}
 `
 
-export function useCurrentUser(
-	userId?: string
-): {
+export function useCurrentUser(userId?: string): {
 	currentUser: User
 	loadCurrentUser: (userId: string) => void
 } {
 	const [currentUser, setCurrentUser] = useRecoilState<User | null>(currentUserState)
-	const [, setUserAuth] = useRecoilState<AuthenticationResponse | null>(userAuthState)
+	// const [, setUserAuth] = useRecoilState<AuthenticationResponse | null>(userAuthState)
 
 	// Handle loading current user
 	const [load] = useLazyQuery(CURRENT_USER, {
 		onCompleted: data => {
+			debugger
 			// Check user permssion here if a user is currently logged in
 			if (data?.user) setCurrentUser(data.user)
 		},
 		onError: error => {
+			debugger
 			console.log('Error loading current user')
 			console.log('Errors on useCurrentUser', error)
-			setCurrentUser(null)
-			setUserAuth(null)
+			// setCurrentUser(null)
+			// setUserAuth(null)
 		}
 	})
 
