@@ -5,22 +5,23 @@
 import { ApolloProvider } from '@apollo/client'
 import { initializeIcons } from '@fluentui/react'
 import type { AppProps } from 'next/app'
-import { useEffect } from 'react'
+import { useEffect, useMemo, memo } from 'react'
 import { createApolloClient } from '~api'
 import { RecoilRoot } from 'recoil'
 
 import '~styles/bootstrap.custom.scss'
 import '~styles/App_reset_styles.scss'
 
-export default function App({ Component, pageProps }: AppProps): JSX.Element {
+const App = memo(function App({ Component, pageProps }: AppProps): JSX.Element {
 	useEffect(() => {
 		initializeIcons()
 	}, [])
+	const apiClient = useMemo(() => createApolloClient(), [])
 
 	return (
 		<>
 			{/* Wrap the page in providers */}
-			<ApolloProvider client={createApolloClient()}>
+			<ApolloProvider client={apiClient}>
 				<RecoilRoot>
 					{/* The Page Component */}
 					<Component className='test' {...pageProps} />{' '}
@@ -28,4 +29,5 @@ export default function App({ Component, pageProps }: AppProps): JSX.Element {
 			</ApolloProvider>
 		</>
 	)
-}
+})
+export default App
